Megashares: regression fixed Vieles von EDT queue auf eigene Threads ausgegliedert! Gui bietet nur eine viel bessre *reaktion* ContainerLinks:optimized Rapidshare: API flood wird erkannt
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@5701 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/plugins/PluginsC.java
@@ -125,7 +125,7 @@ public abstract class PluginsC extends Plugin {
                 next.setName(srcLink.getName());
 
                 if (next.getDownloadSize() < 10) {
-                    next.setDownloadSize((int) srcLink.getDownloadSize());
+                    next.setDownloadSize(srcLink.getDownloadSize());
                 }
 
                 next.getSourcePluginPasswords().addAll(srcLink.getSourcePluginPasswords());
@@ -275,7 +275,6 @@ public abstract class PluginsC extends Plugin {
                 ren = null;
             }
             for (DownloadLink l : failed) {
-
                 l.setContainerIndex(l.getContainerIndex() / 2);
             }
         }
